requirements.txt
numpy<2.3.0
opencv-python
protobuf>=5.26.1,<6.0.0
pydantic
requests
scipy
sqlalchemy
google-cloud-vision
google-ai-generativelanguage